# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  go_boilerplate_internal_modules_role_dto.CreateRoleRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      permissions:
        items:
          type: string
        minItems: 1
        type: array
      slug:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - name
    - permissions
    - slug
    type: object
  go_boilerplate_internal_modules_role_dto.RoleResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      slug:
        type: string
      updated_at:
        type: string
    type: object
  go_boilerplate_internal_modules_role_dto.UpdateRoleRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      permissions:
        items:
          type: string
        minItems: 1
        type: array
    type: object
...
  go_boilerplate_internal_shared_utils.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a robust Go REST API boilerplate using Fiber, GORM, and PostgreSQL.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /oauth/github/callback:
    get:
      description: Handle the callback from GitHub OAuth2.
      parameters:
      - description: Authorization code from GitHub
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "400":
          description: Authentication failed
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      summary: GitHub Callback
      tags:
      - OAuth
...
  /oauth/google/callback:
    get:
      description: Handle the callback from Google OAuth2.
      parameters:
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "400":
          description: Authentication failed
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      summary: Google Callback
      tags:
      - OAuth
  /roles:
    get:
      description: Retrieve a paginated list of all user roles (SuperAdmin only).
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_boilerplate_internal_modules_role_dto.RoleResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new security role (SuperAdmin only).
      parameters:
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/go_boilerplate_internal_modules_role_dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_role_dto.RoleResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Permantently remove a security role (SuperAdmin only).
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      description: Retrieve details of a specific role by its ID (SuperAdmin only).
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role retrieved
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_role_dto.RoleResponse'
              type: object
        "400":
          description: Invalid role ID
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update name or permissions of a security role (SuperAdmin only).
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/go_boilerplate_internal_modules_role_dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_role_dto.RoleResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"